+++ policyengine_us_data/calibration/local_h5/__init__.py
@@ -1,6 +1,10 @@
 """Internal package for the incremental local H5 migration.
 
-Themed subpackages under ``requests``, ``inputs``, ``validation``,
-``results``, and ``partitioning`` are the intended import surfaces.
-This package deliberately does not re-export their contents.
+The initial migration surface is intentionally small:
+
+- ``requests.py`` for typed area request values
+- ``partitioning.py`` for pure worker chunking helpers
+
+Later modules should be added only when they become active runtime
+boundaries rather than speculative contract placeholders.
 """
