--- policyengine_us_data/calibration/local_h5/inputs/__init__.py
@@ -1,9 +1,0 @@
-"""Input-bundle contracts for local H5 publication.
-
-This package owns typed descriptions of the artifact inputs needed for a
-publication run. Future input-resolution helpers should live here.
-"""
-
-from .contracts import PublishingInputBundle
-
-__all__ = ["PublishingInputBundle"]
