+++ tests/unit/calibration/fixtures/test_local_h5_requests.py
@@ -1,9 +1,4 @@
-"""Fixture helpers for ``test_local_h5_contracts.py``.
-
-These helpers load the themed ``local_h5`` subpackages from disk under
-a synthetic package so the repository's heavyweight top-level package
-initializers never run.
-"""
+"""Fixture helpers for ``test_local_h5_requests.py``."""
 
 from __future__ import annotations
 
@@ -16,7 +11,7 @@ __test__ = False
 
 
 def _ensure_package(name: str, path: Path) -> None:
-    """Register a synthetic package so relative imports resolve locally."""
+    """Register a synthetic package so local imports resolve from disk."""
 
     package = sys.modules.get(name)
     if package is None:
@@ -40,8 +35,8 @@ def _load_module(name: str, path: Path):
     return module
 
 
-def load_contracts_exports():
-    """Load the themed local H5 contract packages and return key exports."""
+def load_requests_exports():
+    """Load the local H5 request module under a synthetic package name."""
 
     local_h5_root = (
         Path(__file__).resolve().parents[4]
@@ -49,47 +44,27 @@ def load_contracts_exports():
         / "calibration"
         / "local_h5"
     )
-    package_name = "local_h5_fixture"
+    package_name = "local_h5_requests_fixture"
 
     for name in list(sys.modules):
         if name == package_name or name.startswith(f"{package_name}."):
             sys.modules.pop(name, None)
 
     _ensure_package(package_name, local_h5_root)
-
     requests_module = _load_module(
         f"{package_name}.requests",
-        local_h5_root / "requests" / "__init__.py",
+        local_h5_root / "requests.py",
     )
-    inputs_module = _load_module(
-        f"{package_name}.inputs",
-        local_h5_root / "inputs" / "__init__.py",
-    )
-    validation_module = _load_module(
-        f"{package_name}.validation",
-        local_h5_root / "validation" / "__init__.py",
-    )
-    results_module = _load_module(
-        f"{package_name}.results",
-        local_h5_root / "results" / "__init__.py",
-    )
-    package_module = _load_module(package_name, local_h5_root / "__init__.py")
     return {
-        "module": package_module,
+        "module": requests_module,
         "AreaBuildRequest": requests_module.AreaBuildRequest,
-        "AreaBuildResult": results_module.AreaBuildResult,
         "AreaFilter": requests_module.AreaFilter,
-        "PublishingInputBundle": inputs_module.PublishingInputBundle,
-        "ValidationIssue": validation_module.ValidationIssue,
-        "ValidationPolicy": validation_module.ValidationPolicy,
-        "ValidationResult": validation_module.ValidationResult,
-        "WorkerResult": results_module.WorkerResult,
         "make_national_request": make_national_request,
     }
 
 
 def make_national_request(area_build_request_cls):
-    """Build the canonical national request shape used by current tests."""
+    """Build the canonical national request shape used by request tests."""
 
     return area_build_request_cls(
         area_type="national",
